+++ src/response.rs
@@ -62,7 +62,7 @@ pub struct DealResponse {
   pub proposal_id: String,
   pub freelancer_id: String,
   pub client_id: String,
-  pub price: usize,
+  pub price: f64,
   pub status: String,
   pub address: String,
 }
@@ -73,7 +73,7 @@ pub struct PartialDealResponse {
   pub proposal_id: String,
   pub freelancer_id: String,
   pub client_id: String,
-  pub price: usize,
+  pub price: f64,
   pub status: String,
   pub address: String,
 }
